CA-173605: swap the order of memory observation on host and running domains
This patch simply swaps the order of getting host memory information and
domains memory information during memory planning.

As already stated in the source code comment, we don't currently have a way to
atomically get the current memory layout/consumption of both the host and all
its running domains at the same time. We have to do this one step after
another, which can potentially cause inconsistency due to time gaps. We
accommodate this by adding tolerances in the planning and also expect an
amiable memory allocation speed and an eventual convergence.

This is generally fine in most situations, the worst case that could happen is
that 1) there is intensive domain memory allocation ongoing and 2) the
inconsistency causes us to think that we have _more_ free host memory than we
actually has (as the time of the actual planing), in which case we might be
short of memory by encouraging multiple operations to compete for the same
amount.

By swapping the host memory inspection step to later (than domains memory
inspection), we makes the calculation more conservative in such cases: the
host free_mem observation will be closer to the final memory planning step
hence smaller (in cases of intensive memory allocation) than the other way
around; also a domain's memory_actual observation will be further ahead so
that the estimated memory reservation for that domain (to come from the host
free_mem pool) will come out larger. So, eventually the host's available free
memory will be estimated more conservative than the current implementation, and
therefore safer.

Signed-off-by: Zheng Li <dev@zheng.li>