CA-335725: Calculate a constant boot time host free memory
Free memory of the host is changing when boot and it takes a while to be
constant.

To get a constant value of the boot time host free memory to make xapi able to
calculate the accurate free memory of the host.

1. Add 2 paramters to configure 1) how many same check results mean the free memory
is constant. 2) interval of checks
2. Run the free memory checker in a thread because it usually takes dozens of
seconds.
3. The `get_host_initial_free_memory` API will be blocked until the free memory
is calculated.

Signed-off-by: Yang Qian <yang.qian@citrix.com>